Q7,  Reverse input value in Arrays
Write a program that gets the array size and elements as input from the user. Then,
create another array of the same size as the first array and copy the values of the first
array in reverse order in the second array (means the value on the first index of the first
array should be saved on the last index of the second array, and so on). Then, display the
values of both arrays