Q8 , Sum of 2 arrays In Third
Write a program where you should sum the elements of two same-sized arrays and the
result should be saved in the third array. Then, print the values of all of the three arrays.
The values of the first two arrays may be given fixed (no need to get input) but each value
of the third array should be the sum of respective values from the first and the second
arrays.