Q9 Merge first 2 arrays in third arrays
Write a program that creates two integer arrays of sizes N and M, respectively. These
arrays can be initialized within the code (no need to get values as input from the user).
Then, create a third array of size N+M (i.e., whose size should be equal to the sum of the
sizes of the first and the second array) and copy the elements of both (first and second)
arrays into the third array. Finally, print the values of all the three arrays