Q2 count the  number of  words in sentence
Create a user-defined function that should take one string parameter and it should
return an integer value: the count or number of words in that string. Then, call it in the
main ( ), where you should ask the user to enter any string value (can be an entire line),
pass that value to the function being called, and display the returned value.